+++ zkevm-circuits/src/evm_circuit/execution/chainid.rs
@@ -9,7 +9,7 @@ use crate::{
         },
         witness::{Block, Call, ExecStep, Transaction},
     },
-    table::BlockContextFieldTag,
+    table::block_table::BlockContextFieldTag,
     util::Expr,
 };
 use bus_mapping::evm::OpcodeId;
