+++ zkevm-circuits/src/evm_circuit/execution/end_block.rs
@@ -9,7 +9,7 @@ use crate::{
         },
         witness::{Block, Call, ExecStep, Transaction},
     },
-    table::{CallContextFieldTag, TxContextFieldTag},
+    table::{tx_table::TxContextFieldTag, CallContextFieldTag},
     util::Expr,
 };
 use eth_types::Field;
