+++ zkevm-circuits/src/witness/block.rs
@@ -1,15 +1,11 @@
 use std::collections::HashMap;
 
-use crate::{
-    evm_circuit::{detect_fixed_table_tags, util::rlc, EvmCircuit},
-    table::BlockContextFieldTag,
-};
+use crate::evm_circuit::{detect_fixed_table_tags, EvmCircuit};
 use bus_mapping::{
     circuit_input_builder::{self, CircuitsParams, CopyEvent, ExpEvent},
     Error,
 };
-use eth_types::{Address, Field, ToLittleEndian, ToScalar, Word};
-use halo2_proofs::circuit::Value;
+use eth_types::{Address, Field, Word};
 
 use super::{step::step_convert, tx::tx_convert, Bytecode, ExecStep, RwMap, Transaction};
 
@@ -156,70 +152,6 @@ pub struct BlockContext {
     pub history_hashes: Vec<Word>,
     /// The chain id
     pub chain_id: Word,
-}
-
-impl BlockContext {
-    /// Assignments for block table
-    pub fn table_assignments<F: Field>(&self, randomness: Value<F>) -> Vec<[Value<F>; 3]> {
-        [
-            vec![
-                [
-                    Value::known(F::from(BlockContextFieldTag::Coinbase as u64)),
-                    Value::known(F::zero()),
-                    Value::known(self.coinbase.to_scalar().unwrap()),
-                ],
-                [
-                    Value::known(F::from(BlockContextFieldTag::Timestamp as u64)),
-                    Value::known(F::zero()),
-                    Value::known(self.timestamp.to_scalar().unwrap()),
-                ],
-                [
-                    Value::known(F::from(BlockContextFieldTag::Number as u64)),
-                    Value::known(F::zero()),
-                    Value::known(self.number.to_scalar().unwrap()),
-                ],
-                [
-                    Value::known(F::from(BlockContextFieldTag::Difficulty as u64)),
-                    Value::known(F::zero()),
-                    randomness
-                        .map(|randomness| rlc::value(&self.difficulty.to_le_bytes(), randomness)),
-                ],
-                [
-                    Value::known(F::from(BlockContextFieldTag::GasLimit as u64)),
-                    Value::known(F::zero()),
-                    Value::known(F::from(self.gas_limit)),
-                ],
-                [
-                    Value::known(F::from(BlockContextFieldTag::BaseFee as u64)),
-                    Value::known(F::zero()),
-                    randomness
-                        .map(|randomness| rlc::value(&self.base_fee.to_le_bytes(), randomness)),
-                ],
-                [
-                    Value::known(F::from(BlockContextFieldTag::ChainId as u64)),
-                    Value::known(F::zero()),
-                    randomness
-                        .map(|randomness| rlc::value(&self.chain_id.to_le_bytes(), randomness)),
-                ],
-            ],
-            {
-                let len_history = self.history_hashes.len();
-                self.history_hashes
-                    .iter()
-                    .enumerate()
-                    .map(|(idx, hash)| {
-                        [
-                            Value::known(F::from(BlockContextFieldTag::BlockHash as u64)),
-                            Value::known((self.number - len_history + idx).to_scalar().unwrap()),
-                            randomness
-                                .map(|randomness| rlc::value(&hash.to_le_bytes(), randomness)),
-                        ]
-                    })
-                    .collect()
-            },
-        ]
-        .concat()
-    }
 }
 
 impl From<&circuit_input_builder::Block> for BlockContext {
