+++ zkevm-circuits/src/witness/tx.rs
@@ -2,7 +2,7 @@ use bus_mapping::circuit_input_builder;
 use eth_types::{Address, Field, ToLittleEndian, ToScalar, ToWord, Word};
 use halo2_proofs::circuit::Value;
 
-use crate::{evm_circuit::util::rlc, table::TxContextFieldTag, util::Challenges};
+use crate::{evm_circuit::util::rlc, table::tx_table::TxContextFieldTag, util::Challenges};
 
 use super::{step::step_convert, Call, ExecStep};
